refactor: file structure changes
