+++ src/lambda/expression.rs
@@ -112,54 +112,3 @@ impl Eval for Expression {
     })
   }
 }
-
-pub fn get_path_for_const_value_owned(path: &[impl AsRef<str>], mut const_value: ConstValue) -> Option<ConstValue> {
-  for path in path.iter() {
-    const_value = match const_value {
-      ConstValue::Object(mut obj) => obj.remove(path.as_ref())?,
-      _ => None?,
-    }
-  }
-
-  Some(const_value)
-}
-
-pub fn get_path_for_const_value_ref<'a>(
-  path: &[impl AsRef<str>],
-  mut const_value: &'a ConstValue,
-) -> Option<&'a ConstValue> {
-  for path in path.iter() {
-    const_value = match const_value {
-      ConstValue::Object(ref obj) => obj.get(path.as_ref())?,
-      _ => None?,
-    }
-  }
-
-  Some(const_value)
-}
-
-#[cfg(test)]
-mod tests {
-  use async_graphql::{Name, Number, Value};
-  use hyper::body::Bytes;
-  use indexmap::IndexMap;
-
-  use crate::lambda::is_truthy;
-
-  #[test]
-  fn test_is_truthy() {
-    assert!(is_truthy(Value::Enum(Name::new("EXAMPLE"))));
-    assert!(is_truthy(Value::List(vec![])));
-    assert!(is_truthy(Value::Object(IndexMap::default())));
-    assert!(is_truthy(Value::String("Hello".to_string())));
-    assert!(is_truthy(Value::Boolean(true)));
-    assert!(is_truthy(Value::Number(Number::from(1))));
-    assert!(is_truthy(Value::Binary(Bytes::from_static(&[0, 1, 2]))));
-
-    assert!(!is_truthy(Value::Null));
-    assert!(!is_truthy(Value::String("".to_string())));
-    assert!(!is_truthy(Value::Boolean(false)));
-    assert!(!is_truthy(Value::Number(Number::from(0))));
-    assert!(!is_truthy(Value::Binary(Bytes::default())));
-  }
-}
