+++ src/lambda/mod.rs
@@ -2,7 +2,6 @@ mod eval;
 mod evaluation_context;
 mod expression;
 mod graphql_operation_context;
-mod helpers;
 mod io;
 mod lambda;
 mod list;
@@ -15,7 +14,6 @@ pub use eval::*;
 pub use evaluation_context::EvaluationContext;
 pub(crate) use expression::*;
 pub use graphql_operation_context::GraphQLOperationContext;
-pub use helpers::*;
 pub use io::*;
 pub use lambda::Lambda;
 pub use list::*;
